Filling in direct packages references in a more direct and reliable way
In the future we may be able to get the direct reference information from the project asset file.
But until that is available, using the dotnet command allows us get the information in a more reliable way.

Previously, if any direct dependency was also a transitive dependency, then it was
only treated as a transitive dependency (being removed from the top level component).

Signed-off-by: Riley Huddleston <rilehudd@selinc.com>