feat: add evidence collect option
The license evidence collection may be used to collect license information from shipped files, like LICENSE.txt. This is particularly useful for packages, which have no license id provided, but rather information is provided in a file. Also even when the license id or name is known, it still might be a good idea to have the license information from the time of BOM creation.

The default mode is `None` which means no license evidence will be collected. The other options are `All` which collects all license evidence, even when the license id is known. Lastly, `Unknown` Collect license text only for components which have unknown license. This avoids collecting all license texts for the case when license text can be obtained otherwise (like MIT) and therefore reduces the BOM size. In contrast to the "All" mode, this mode will put license text into license block directly instead of evidence part.

Signed-off-by: Roman Malytskyy <malytskyy@users.noreply.github.com>